Always expand the definitions and parameters
- Fixes recursive expansion issues
